requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
braindecode>=0.8.0
mne>=1.3.0
numpy
pandas
matplotlib
seaborn>=0.12.0
scikit-learn
moabb>=1.0.0
tqdm
pyyaml
torchmetrics>=1.0.0
einops>=0.6.0
hydra-joblib-launcher>=1.2.0